+++ NEWS.md
@@ -3,6 +3,10 @@
 
 Development version
 
+`google_distance()` implemented. Returns a distance matrix from Google Distance API call.
+
+`get_route()` is now deprecated. Use `google_directions()` instead.
+
 New arguments for `get_route()` 
 
 * `waypoints`